requirements.txt
django-environ
Django>=4.0.0
djangorestframework
django-background-tasks
